Display PG/VG/Nicotine in Flavor View. Manually added dummy flavors to flavors list before passing to table. These flavors can be identified by their manufacturer (empty string). Minor cleanup of several files.

+++ src/main/java/zbb/entities/Flavor.java
@@ -22,7 +22,7 @@ public class Flavor implements Serializable{
 		this(manf, name, -1);
 	}
 
-	public Flavor(@NotNull String manf, @NotNull String name, int amount) {
+	public Flavor(@NotNull String manf, @NotNull String name, double amount) {
 		this.manufacturer = manf;
 		this.name = name;
 		if (amount > 0) {
